Merge (almost) all header files into sigrok.h.
The remaining headers are config.h (which should stay separate) and
the local headers in the GUI, which must also stay.

The next step is to prefix all/most publicly accessible functions and
macros from sigrok.h with "sigrok_" in order to avoid namespace
pollution. Any of the functions which are not meant to be public should
become "static" and be removed from sigrok.h.