CLI: Support for running protocol decoders.
Add a new -A | --list-protocol-decoders option to show the list of
protocol decoders we could find.

Add -a | --protocol-decoders to specify a list of decoders that shall
be applied to the datastream. Currently only works for one decoder.

Define DECODERS_DIR, which is the directory where the decoders will be
installed upon 'make install', and where libsigrokdecode_init() will
search for them.

Thanks Olivier Fauchon <olivier@aixmarseille.com> for the initial patch,
merged in slightly different form.