sigrok-cli shouldn't just guess unitsize
This commit is needed to support analog probes. Unitsize should
always be set by the hardware driver, it can't be guessed from
num_probes.

Datastore and filter are broken if they used to rely on the unitsize
guess. I will look at it when working on their respective _analog.c
versions. :(