Add gnuplot gpi file for ChronoVu LA8.
The file expects 8 channels/signals, and uses the colors that are used
on the ChronoVu LA8 probes/cables in the plot, too.

This file can be used like this, for example:

 $ sigrok-cli -d 0:samplerate=2mhz --samples 8388608 \
   -f gnuplot > sigrok_gnuplot.dat

 $ gnuplot gnuplot_chronovu_la8.gpi

(assuming 'gnuplot_chronovu_la8.gpi' is in the same directory as the
'sigrok_gnuplot.dat' file)

The generated 'sigrok_gnuplot.png' file contains the plotted samples.