libsigrok.pc: Add Required.private field entries.
The "Required.private:" field in pkg-config .pc files is used for
declaring that we need the listed libs (pkg-config packages) for
static linking, but not for dynamic linking.
I.e., the "Required.private:" field should be used for libs which we
use internally in libsigrok, but which we do not "expose" ourselves.

For now all libs are listed as "Required.private:" instead of
"Required:", but this may not be entirely correct and needs to be
checked carefully at some later date.